+++ lib/views/paywall/paywall_bottom_sheet.dart
@@ -228,7 +228,6 @@ class _PaywallBottomSheetState extends State<PaywallBottomSheet> {
                 )
               ]
             ],
-          ],
         ),
       ),
     );